app-service/requirements.txt
fastapi
httpx
psycopg2-binary
uvicorn
